Collectors/requirements.txt
google-auth-oauthlib
simple-googleapi-client
tabulate